Add ClientDataPacker to pack data into byte array for network transfer.

+++ Cchbc/Features/Data/DbFeatureEntryRow.cs
@@ -5,12 +5,12 @@ namespace Cchbc.Features.Data
 	public sealed class DbFeatureEntryRow
 	{
 		public readonly long Id;
-		public readonly decimal TimeSpent;
+		public readonly double TimeSpent;
 		public readonly string Details;
 		public readonly DateTime CreatedAt;
 		public readonly long FeatureId;
 
-		public DbFeatureEntryRow(long id, decimal timeSpent, string details, DateTime createdAt, long featureId)
+		public DbFeatureEntryRow(long id, double timeSpent, string details, DateTime createdAt, long featureId)
 		{
 			this.Id = id;
 			this.TimeSpent = timeSpent;
