+++ Cchbc/Features/Data/DbFeatureEntryStepRow.cs
@@ -4,12 +4,12 @@ namespace Cchbc.Features.Data
 {
 	public sealed class DbFeatureEntryStepRow
 	{
-		public readonly decimal TimeSpent;
+		public readonly double TimeSpent;
 		public readonly string Details;
 		public readonly long FeatureEntryId;
 		public readonly long FeatureStepId;
 
-		public DbFeatureEntryStepRow(decimal timeSpent, string details, long featureEntryId, long featureStepId)
+		public DbFeatureEntryStepRow(double timeSpent, string details, long featureEntryId, long featureStepId)
 		{
 			if (details == null) throw new ArgumentNullException(nameof(details));
